tests/submission/MOSQUITO/pattern_attribute.html

<!DOCTYPE html>
<html>

  <head>
    <title>Pattern Attribute</title>
    <meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	<meta name="author" lang="fr" content="Fabrice Clari, Dimitri Bocquet">
    <script src="http://w3c-test.org/resources/testharness.js"></script>
    <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
  </head>

  <body>

  
  		<h1>Pattern Attribute</h1>
  		<div style="display: none">
  		<input pattern="[a-z]{3}" value="abcd" title="three letters max"/>
		</div>
		
    <div id="log">
    </div>
	
	<script type="text/javascript">


		test(function() {assert_equals(document.getElementsByTagName("input")[0].getAttribute("pattern"), "[a-z]{3}")}, "pattern attribute support on input element");

	</script>
	
  </body>
  
</html>
